Run clippy everywhere (#372)
Instead of whitelisting a few packages to run clippy on, we just
blacklist the two where it's not expected to work.

Based on https://github.com/scroll-tech/ceno/pull/371. So when that PR
merges, this PR's diff will become smaller. But we might still have to
break this one down into multiple PRs.

+++ gkr/src/verifier.rs
@@ -49,7 +49,7 @@ impl<E: ExtensionField> IOPVerifierState<E> {
             let timer = start_timer!(|| format!("Verify layer {}", layer_id));
             verifier_state.layer_id = layer_id as LayerId;
 
-            let layer = &circuit.layers[layer_id as usize];
+            let layer = &circuit.layers[layer_id];
             for (step, step_proof) in izip!(layer.sumcheck_steps.iter(), &mut sumcheck_proofs_iter)
             {
                 match step {
