+++ singer-utils/src/structs.rs
@@ -1,5 +1,4 @@
-use ff_ext::ExtensionField;
-use simple_frontend::structs::{ChallengeId, ExtCellId};
+use simple_frontend::structs::ChallengeId;
 use strum_macros::EnumIter;
 use uint::UInt;
 
@@ -41,4 +40,4 @@ pub struct ChipChallenges {
 pub type UInt64 = UInt<64, VALUE_BIT_WIDTH>;
 pub type PCUInt = UInt64;
 pub type TSUInt = UInt<48, 48>;
-pub type StackUInt = UInt<{ EVM_STACK_BIT_WIDTH as usize }, { VALUE_BIT_WIDTH as usize }>;
+pub type StackUInt = UInt<{ EVM_STACK_BIT_WIDTH }, { VALUE_BIT_WIDTH }>;
